Collector (#1)
* started constants, made enum

* added neutral mode and inverted mode value

* fixed small line mistakes

* fixed TalonSRX config

* added voltage compensation

* removed extra new line

* removed extra new line in subsystem

* updated code acording to review

* removed extra new line

* added motor to subsystem

* added note

* added stop and atempted setCollectorStateCommand

btw, I doubt motor.setIntegralAccumulator is right but it's my best guess so far.

* finished subsystem

* fixed

* added javadocs

* removed extra new lines

* fixed javadocs

* updated command

* added currentWatcher

I attempted to add the currentWatcher to the collector. I also chenged the currentSupplier in the currentWatcher class from int to double because that's the only way it worked with the motor. I also just randomly guessed numbers for the max time and maxCurrent as I have no idea what to put there

* fixed implementation of CurrentWatcher

* updated numbers

* corrected implementation of currentWatcher

* made startCurrentWatcher

* changed neutralmode to coast

+++ src/main/java/frc/trigon/robot/utilities/CurrentWatcher.java
@@ -2,7 +2,6 @@ package frc.trigon.robot.utilities;
 
 import edu.wpi.first.wpilibj.Notifier;
 import edu.wpi.first.wpilibj.Timer;
-
 import java.util.function.Supplier;
 
 /**
@@ -10,7 +9,7 @@ import java.util.function.Supplier;
  */
 public class CurrentWatcher {
     private final Runnable runnable;
-    private final Supplier<Integer> currentSupplier;
+    private final Supplier<Double> currentSupplier;
     private final int maxCurrent;
     private final double maxTime;
     private double belowCurrentTimeStamp;
@@ -23,7 +22,7 @@ public class CurrentWatcher {
      * @param maxCurrent      the current limit
      * @param maxTime         the time needed for the current to exceed its limit, in order to run the runnable
      */
-    public CurrentWatcher(Runnable runnable, Supplier<Integer> currentSupplier, int maxCurrent, double maxTime) {
+    public CurrentWatcher(Runnable runnable, Supplier<Double> currentSupplier, int maxCurrent, double maxTime) {
         this.runnable = runnable;
         this.currentSupplier = currentSupplier;
         this.maxCurrent = maxCurrent;
